Move ShipEngine-specific RestClient into ShipEngine directory
It turns out this client is more specific to ShipEngine than I thought.
Rename it so it lives in the ShipEngine directory.
